+++ src/apps/chifra/pkg/pinlib/manifest/manifest.go
@@ -1,11 +1,5 @@
 package manifest
 
-import (
-	"os"
-
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/config"
-)
-
 type IpfsHash = string
 
 type PinDescriptor struct {
@@ -24,14 +18,4 @@ type Manifest struct {
 	PreviousBlockRange ManifestRange   `json:"prevBlockRange"`
 	NewPins            []PinDescriptor `json:"newPins"`
 	PreviousPins       []PinDescriptor `json:"prevPins"`
-}
-
-// FromLocalFile loads the manifest saved in ConfigPath
-func FromLocalFile() (*Manifest, error) {
-	file, err := os.Open(config.GetConfigPath("manifest/manifest.txt"))
-	if err != nil {
-		return nil, err
-	}
-
-	return ReadTabManifest(file)
 }
